Restructure firebase configuration
